add ability to reset rotation via signal + gamepad
The gamepad change isn't the best as it requires the window to be
focused, but it allows to use a wireless gamepad to reset the rotation
and position like using the w key on the keyboard. I currently hardcoded
the reset button to 6 on a joystick, which was equal to the select
button on my steam controller using sc-controller simulating an xbox
controller. Maybe the controller update isn't quite fit to stay and
should use the VR controllers instead & work without focus.

The more interesting part of this patch is adding signal handling of
SIGUSR1 or SIGUSR2 to do the same thing. This allows to simply run a
command like `killall -USR1 vr-video-player` to reset the rotation and
position even without window focus. This is especially useful for
example for adding a custom command action to KDE connect so you can
reset the VR view from your phone without needing to have any special
software.

However I believe even this signal handling should eventually be
replaced with some more advanced API like dbus or a unix socket to also
allow for more advanced features like changing tracked windows on the
fly.